requirements.txt
flask
python-telegram-bot
requests
gunicorn
pandas
flask-sqlalchemy